+++ samples/main.py
@@ -4,7 +4,8 @@ import os
 from typing import Any, Dict
 import threading
 
-from utils.audio_processing import AudioPlayer, AudioCapture, AudioCaptureEventHandler
+from utils.audio_playback import AudioPlayer
+from utils.audio_capture import AudioCapture, AudioCaptureEventHandler
 from realtime_ai.realtime_ai_client import RealtimeAIClient
 from realtime_ai.models.realtime_ai_options import RealtimeAIOptions
 from realtime_ai.models.audio_stream_options import AudioStreamOptions
@@ -12,15 +13,21 @@ from realtime_ai.realtime_ai_event_handler import RealtimeAIEventHandler
 from realtime_ai.models.realtime_ai_events import *
 
 # Set up logging
-logging.getLogger("audio_processing").setLevel(logging.DEBUG)
+# Configure logging
+logging.basicConfig(
+    level=logging.DEBUG,
+    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
+    handlers=[logging.StreamHandler()]  # Streaming logs to the console
+)
+
+# Specific loggers for mentioned packages
+logging.getLogger("utils.audio_playback").setLevel(logging.DEBUG)
+logging.getLogger("utils.audio_capture").setLevel(logging.ERROR)
+logging.getLogger("utils.vad").setLevel(logging.ERROR)
 logging.getLogger("realtime_ai").setLevel(logging.ERROR)
 
+# Root logger for general logging
 logger = logging.getLogger()
-if not logger.hasHandlers():
-    logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(levelname)s - %(message)s')
-else:
-    for handler in logger.handlers:
-        handler.setFormatter(logging.Formatter('%(asctime)s - %(levelname)s - %(message)s'))
 
 
 class MyAudioCaptureEventHandler(AudioCaptureEventHandler):
@@ -65,6 +72,9 @@ class MyAudioCaptureEventHandler(AudioCaptureEventHandler):
             current_audio_content_index = self.event_handler.get_current_audio_content_id()
             logger.info(f"Truncate the current audio, current item ID: {current_item_id}, current audio content index: {current_audio_content_index}")
             self.client.truncate_response(item_id=current_item_id, content_index=current_audio_content_index, audio_end_ms=1000)
+
+            # Restart the audio player
+            self.event_handler.audio_player.drain_and_restart(clear_buffer=True)
         else:
             logger.info("Assistant is not speaking, cancelling response is not required.")
             self.cancelled = False
@@ -81,12 +91,16 @@ class MyAudioCaptureEventHandler(AudioCaptureEventHandler):
 class MyRealtimeEventHandler(RealtimeAIEventHandler):
     def __init__(self, audio_player: AudioPlayer):
         super().__init__()
-        self.audio_player = audio_player
+        self._audio_player = audio_player
         self.audio_buffer = []
         self.client = None
         self.current_item_id = None
         self.current_audio_content_index = None
         self._is_audio_playing = False
+
+    @property
+    def audio_player(self):
+        return self._audio_player
 
     def get_current_conversation_item_id(self):
         return self.current_item_id
@@ -139,7 +153,7 @@ class MyRealtimeEventHandler(RealtimeAIEventHandler):
         logger.info(f"Audio done for response ID {event.response_id}, item ID {event.item_id}")
         if self.client:
             self._is_audio_playing = False
-        self.audio_player.drain_and_restart()
+        self._audio_player.drain_and_restart()
 
     def on_response_audio_transcript_done(self, event: ResponseAudioTranscriptDone):
         logger.info(f"Audio transcript done: '{event.transcript}' for response ID {event.response_id}")
@@ -182,7 +196,7 @@ class MyRealtimeEventHandler(RealtimeAIEventHandler):
         if delta_audio:
             try:
                 audio_bytes = base64.b64decode(delta_audio)
-                self.audio_player.enqueue_audio_data(audio_bytes)
+                self._audio_player.enqueue_audio_data(audio_bytes)
                 self._is_audio_playing = True
             except base64.binascii.Error as e:
                 logger.error(f"Failed to decode audio delta: {e}")
@@ -239,7 +253,7 @@ def main():
         )
 
         # Initialize AudioPlayer
-        audio_player = AudioPlayer(min_buffer_fill=3)
+        audio_player = AudioPlayer(min_buffer_fill=3, enable_wave_capture=True)
 
         # Initialize RealtimeAIClient with MyRealtimeEventHandler to handle events
         event_handler = MyRealtimeEventHandler(audio_player=audio_player)
